requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cssselect2==0.4.1
defusedxml==0.6.0
discord.py==1.6.0
idna==2.10
multidict==5.1.0
Pillow==8.1.0
pycparser==2.20
python-dotenv==0.15.0
requests==2.25.1
tinycss2==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1
yarl==1.6.3
